% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileExist.R
\name{dirExistVerbose}
\alias{dirExistVerbose}
\title{Extended directory exist check}
\usage{
dirExistVerbose(dirName, actionFun)
}
\arguments{
\item{actionFun}{A function to run if folder doesn't exist}
}
\value{

}
\description{
Extended directory exist check
}
