% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileExist.R
\name{fileExistVerbose}
\alias{fileExistVerbose}
\title{Extended file exist check}
\usage{
fileExistVerbose(fileName, actionFun)
}
\arguments{
\item{actionFun}{A function to run if file doesn't exist}
}
\value{

}
\description{
Extended file exist check
}
