% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotatePolarCoords.R
\name{rotatePolarCoords}
\alias{rotatePolarCoords}
\title{rotatePolarCoords}
\usage{
rotatePolarCoords(zenith, azimuth, rotateVal)
}
\arguments{
\item{rotateVal}{}
}
\value{
`data.frame`
}
\description{
rotatePolarCoords
}
